pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vulnhawk"
version = "0.1.0"
description = "AI-powered code security scanner that finds vulnerabilities Semgrep and CodeQL miss"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Moamen Basel" }]
keywords = [
    "security",
    "scanner",
    "vulnerability",
    "ai",
    "llm",
    "code-analysis",
    "sast",
    "appsec",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "anthropic>=0.40.0",
    "openai>=1.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "pathspec>=0.12.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[project.scripts]
vulnhawk = "vulnhawk.cli:main"

[project.urls]
Homepage = "https://github.com/momenbasel/vulnhawk"
Repository = "https://github.com/momenbasel/vulnhawk"
Issues = "https://github.com/momenbasel/vulnhawk/issues"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/fixtures/*" = ["E501"]
"vulnhawk/rules/prompts.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
